refactor: move package to pkg
BREAKING CHANGE: rename github.com/nais/naistrix/output to
github.com/nais/naistrix/pkg/output
