--- examples/confirmation/README.md
@@ -1,3 +1,0 @@
-# Conformation dialog
-
-An example showing how to ask the user for confirmation before doing something.
